github key 변경 후 커밋 테스트

+++ src/main/java/fc/algorithm/datastructure/AlpaCount.java
@@ -10,7 +10,7 @@ public class AlpaCount {
         int[] alpa = new int[26];
 
         for (int i = 0; i < s.length(); i++) {
-        	int x = (int)s.substring(i, i+1).charAt(0);
+        	int x = s.substring(i, i+1).charAt(0);
         	for (int j=97; j<=122; j++) {
         		if (x == j) {
         			alpa[j - 97] = alpa[j - 97] + 1;
